[UHYU-211] chore: age type Byte -> Integer 변경

+++ src/test/java/com/ureca/uhyu/domain/user/service/UserServiceTest.java
@@ -394,7 +394,7 @@ class UserServiceTest {
                 .userName("홍길동")
                 .kakaoId(456465L)
                 .email("asdad@kakao.com")
-                .age((byte) 32)
+                .age((Integer) 32)
                 .gender(Gender.MALE)
                 .role(UserRole.TMP_USER)
                 .status(Status.ACTIVE)
